Add Save method to StepsRepository
The StepsRepository interface has been expanded to include a Save method that saves a step by sequence within a database. This update also includes a corresponding handler in main.go. Additionally, a redundant function was removed from pipeline-handler.go and the 'Sequence' parameter in types/pipeline.go was updated to use int64.

+++ web/types/pipeline.go
@@ -18,7 +18,7 @@ type Parameters struct {
 
 type Step struct {
 	ID         uuid.UUID `gorm:"primaryKey"`
-	Sequence   int
+	Sequence   int64
 	Action     string
 	Name       string
 	PipelineID uuid.UUID
